+++ crates/fluvio/src/sockets.rs
@@ -50,7 +50,11 @@ impl VersionedSocket {
         // now get versions
         // Query for API versions
 
-        let mut req_msg = RequestMessage::new_request(ApiVersionsRequest::default());
+        let mut req_msg = RequestMessage::new_request(ApiVersionsRequest {
+            client_version: crate::built_info::PKG_VERSION.into(),
+            client_os: crate::built_info::CFG_OS.into(),
+            client_arch: crate::built_info::CFG_TARGET_ARCH.into(),
+        });
         req_msg.get_mut_header().set_client_id(&config.client_id);
 
         debug!("querying versions");
@@ -186,10 +190,10 @@ impl Versions {
     }
 
     /// Given an API key, it returns max_version. None if not found
-    pub fn lookup_version(&self, api_key: u16) -> Option<i16> {
+    pub fn lookup_version(&self, api_key: u16, client_version: i16) -> Option<i16> {
         for version in &self.api_versions {
             if version.api_key == api_key as i16 {
-                return Some(version.max_version);
+                return Some(std::cmp::min(version.max_version, client_version));
             }
         }
         None
@@ -233,7 +237,11 @@ impl VersionedSerialSocket {
     where
         R: Request + Send + Sync,
     {
-        let req_msg = self.new_request(request, self.versions.lookup_version(R::API_KEY));
+        let req_msg = self.new_request(
+            request,
+            self.versions
+                .lookup_version(R::API_KEY, R::DEFAULT_API_VERSION),
+        );
 
         // send request & save response
         self.socket.send_and_receive(req_msg).await
@@ -260,5 +268,35 @@ impl VersionedSerialSocket {
 impl SerialFrame for VersionedSerialSocket {
     fn config(&self) -> &ClientConfig {
         &self.config
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use dataplane::versions::ApiVersionKey;
+
+    use super::ApiVersionsResponse;
+    use super::Versions;
+
+    #[test]
+    fn test_version_lookup() {
+        let mut response = ApiVersionsResponse::default();
+
+        response.api_keys.push(ApiVersionKey {
+            api_key: 1000,
+            min_version: 0,
+            max_version: 10,
+        });
+
+        let versions = Versions::new(response);
+
+        // None if api_key not found
+        assert_eq!(versions.lookup_version(0, 10), None);
+
+        // Must use max version of the client
+        (0..10).for_each(|i| assert_eq!(versions.lookup_version(1000, i), Some(i)));
+
+        // Since max_version of the client is larger than the max_version of the server, should use the max_version of the server
+        (10..12).for_each(|i| assert_eq!(versions.lookup_version(1000, i), Some(10)));
     }
 }